+++ Sunningdale/MHSystems/app/src/main/java/com/mh/systems/sunningdale/activites/ClubNewsActivity.java
@@ -6,7 +6,10 @@ import android.os.Bundle;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.support.v7.widget.Toolbar;
 import android.util.Log;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
@@ -46,6 +49,9 @@ public class ClubNewsActivity extends BaseActivity {
     /*********************************
      * INSTANCES OF CLASSES
      *******************************/
+    @Bind(R.id.tbClubNews)
+    Toolbar tbClubNews;
+
     @Bind(R.id.llHomeIcon)
     LinearLayout llHomeIcon;
 
@@ -53,11 +59,6 @@ public class ClubNewsActivity extends BaseActivity {
     RecyclerView rvClubNewsList;
 
     ClubNewsSwipeAdapter clubNewsSwipeAdapter;
-
-//    ClubNewsAPI clubNewsAPI;
-//    AJsonParamsClubNews aJsonParamsClubNews;
-
-    ClubNewsItems clubNewsItems;
 
     ClubNewsDetailAPI clubNewsDetailAPI;
     AJsonParamsClubNewsDetail aJsonParamsClubNewsDetail;
@@ -71,6 +72,8 @@ public class ClubNewsActivity extends BaseActivity {
     AJsonParamsClubNewsThumbnail aJsonParamsClubNewsThumbnail;
 
     ClubNewsThumbnailResponse clubNewsThumbnailResponse;
+
+    Intent intent;
 
       /* ++ INTERNET CONNECTION PARAMETERS ++ */
 
@@ -91,7 +94,6 @@ public class ClubNewsActivity extends BaseActivity {
     /*********************************
      * INSTANCES OF LOCAL DATA TYPE
      *******************************/
-    //ArrayList<ClubNewsData> clubNewsDataArrayList = new ArrayList<>();
     ArrayList<ClubNewsThumbnailData> clubNewsThumbnailList = new ArrayList<>();
 
     private Boolean isDelete = true, isRead = true;
@@ -116,26 +118,25 @@ public class ClubNewsActivity extends BaseActivity {
         //Initialize view resources.
         ButterKnife.bind(this);
 
+        if (tbClubNews != null) {
+            setSupportActionBar(tbClubNews);
+            getSupportActionBar().setDisplayShowTitleEnabled(false);
+        }
+
         //Clear Push ArrayList.
         PushNotificationService.strArrList.clear();
 
         // Layout Managers:
         rvClubNewsList.setLayoutManager(new LinearLayoutManager(this));
-        // Item Decorator:
-        //rvClubNewsList.addItemDecoration(new DividerItemDecoration(getResources().getDrawable(R.drawable.divider)));
-        // mRecyclerView.setItemAnimator(new FadeInLeftAnimator());
 
         /**
          *  Check internet connection before hitting server request.
          */
         if (isOnline(this)) {
             showNoInternetView(inc_message_view, ivMessageSymbol, tvMessageTitle, tvMessageDesc, true);
-            // inc_message_view.setVisibility(View.GONE);
             requestClubNews();
         } else {
             showNoInternetView(inc_message_view, ivMessageSymbol, tvMessageTitle, tvMessageDesc, false);
-            // inc_message_view.setVisibility(View.VISIBLE);
-            //showAlertMessage(getResources().getString(R.string.error_no_internet));
             hideProgress();
         }
 
@@ -146,21 +147,41 @@ public class ClubNewsActivity extends BaseActivity {
         llHomeIcon.setOnClickListener(mHomePressListener);
     }
 
-   /* @Override
-    protected void onResume() {
-        super.onResume();
-
-        clubNewsSwipeAdapter.notifyDataSetChanged();
-    }*/
-
     @Override
     public void onBackPressed() {
         super.onBackPressed();
 
-        Intent intent = new Intent(ClubNewsActivity.this, DashboardActivity.class);
+        intent = new Intent(ClubNewsActivity.this, DashboardActivity.class);
         intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_NEW_TASK);
         startActivity(intent);
         finish();
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_news_webcam, menu);
+        return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+        switch (item.getItemId()) {
+            case android.R.id.home:
+                finish();
+                break;
+
+            case R.id.action__webcam:
+                intent = new Intent(ClubNewsActivity.this, ClubNewsWebCamActivity.class);
+                startActivity(intent);
+                finish();
+
+                break;
+
+            default:
+                break;
+        }
+        return true;
     }
 
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
@@ -182,108 +203,6 @@ public class ClubNewsActivity extends BaseActivity {
             clubNewsSwipeAdapter.notifyDataSetChanged();
         }
     }
-
-    /* @Override
-    protected void onResume() {
-        super.onResume();
-
-        *//**
-     *  Check internet connection before hitting server request.
-     *//*
-        if (isOnline(this)) {
-            showNoInternetView(inc_message_view, ivMessageSymbol, tvMessageTitle, tvMessageDesc, true);
-            // inc_message_view.setVisibility(View.GONE);
-            requestClubNews();
-        } else {
-            showNoInternetView(inc_message_view, ivMessageSymbol, tvMessageTitle, tvMessageDesc, false);
-            // inc_message_view.setVisibility(View.VISIBLE);
-            //showAlertMessage(getResources().getString(R.string.error_no_internet));
-            hideProgress();
-        }
-    }*/
-
-    /**
-     * Implement a method to hit News web service to get response.
-     */
-  /*  public void requestClubNews() {
-
-        showPleaseWait("Please wait...");
-
-        aJsonParamsClubNews = new AJsonParamsClubNews();
-        aJsonParamsClubNews.setLoginMemberId(getMemberId());
-
-        clubNewsAPI = new ClubNewsAPI(getClientId(), "GETCLUBNEWS", aJsonParamsClubNews, ApplicationGlobal.TAG_GCLUB_WEBSERVICES);
-
-        //Creating a rest adapter
-        RestAdapter adapter = new RestAdapter.Builder()
-                .setEndpoint(WebAPI.API_BASE_URL)
-                .build();
-
-        //Creating an object of our api interface
-        WebServiceMethods api = adapter.create(WebServiceMethods.class);
-
-        //Defining the method
-        api.getClubNews(clubNewsAPI, new Callback<JsonObject>() {
-            @Override
-            public void success(JsonObject jsonObject, retrofit.client.Response response) {
-
-                updateSuccessResponse(jsonObject);
-            }
-
-            @Override
-            public void failure(RetrofitError error) {
-                //you can handle the errors here
-                Log.e(LOG_TAG, "RetrofitError : " + error);
-                hideProgress();
-                showAlertMessage("" + getResources().getString(R.string.error_please_retry));
-            }
-        });
-    }*/
-
-    /**
-     * Implements this method to UPDATE the data from webservice in
-     * COURSE DIARY list if get SUCCESS.
-     */
-    /*private void updateSuccessResponse(JsonObject jsonObject) {
-        Log.e(LOG_TAG, "SUCCESS RESULT : " + jsonObject.toString());
-
-        Type type = new TypeToken<ClubNewsItems>() {
-        }.getType();
-        clubNewsItems = new com.newrelic.com.google.gson.Gson().fromJson(jsonObject.toString(), type);
-
-        //Clear Old data from Array-list.
-        clubNewsDataArrayList.clear();
-
-        try {
-            *//**
-     *  Check "Result" 1 or 0. If 1, means data received successfully.
-     *//*
-            if (clubNewsItems.getMessage().equalsIgnoreCase("Success")) {
-
-                //Take backup of List before changing to record.
-                clubNewsDataArrayList.addAll(clubNewsItems.getData());
-
-                if (clubNewsDataArrayList.size() == 0) {
-                    clubNewsDataArrayList.clear();
-                    clubNewsSwipeAdapter.notifyDataSetChanged();
-                    showNoCourseView(false);
-                } else {
-                    showNoCourseView(true);
-
-                    //Set Adapter.
-                    clubNewsSwipeAdapter.notifyDataSetChanged();
-                }
-            } else {
-                showNoCourseView(false);
-            }
-        } catch (Exception e) {
-            Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
-        }
-
-        //Dismiss progress dialog.
-        hideProgress();
-    }*/
 
     /**
      * Implements a method to show 'NO COURSE' view and hide it at least one Course event.